pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dewi"
version = "0.1.0"
description = "DEWI: A Design for an Entropy-Weighted Index for Text+Image Corpora"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.21.0",
    "torch>=1.9.0",
    "transformers>=4.15.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",  # or faiss-gpu if CUDA is available
    "Pillow>=8.4.0",
    "tqdm>=4.62.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "hnswlib>=0.7.0",
    "dataclasses>=0.6;python_version<'3.7'"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.12b0",
    "isort>=5.10.1",
    "mypy>=0.910",
    "pylint>=2.12.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
